move makeParameterFile to ifm_sentinel.py to avoid circular import
