+++ hyp3_insar_gamma/ifm_sentinel.py
@@ -2,8 +2,10 @@
 
 import argparse
 import datetime
+import glob
 import logging
 import os
+import re
 import shutil
 import sys
 
@@ -16,7 +18,6 @@ from hyp3_insar_gamma.create_metadata_insar_gamma import create_readme_file
 from hyp3_insar_gamma.getDemFileGamma import getDemFileGamma
 from hyp3_insar_gamma.interf_pwr_s1_lt_tops_proc import interf_pwr_s1_lt_tops_proc
 from hyp3_insar_gamma.par_s1_slc import par_s1_slc
-from hyp3_insar_gamma.stack_sentinel import makeParameterFile
 from hyp3_insar_gamma.unwrapping_geocoding import unwrapping_geocoding
 
 # FIXME: refactor to eliminate globals
@@ -209,6 +210,77 @@ def move_output_files(outdir, output, reference, prod_dir, long_output, los_flag
 
     makeAsfBrowse("{}.adf.unw.geo.bmp.tif".format(os.path.join(outdir, output)),
                   "{}_unw_phase".format(os.path.join(prod_dir, long_output)))
+
+
+def makeParameterFile(mydir, alooks, rlooks, dem_source):
+    res = 20 * int(alooks)
+
+    reference_date = mydir[:15]
+    secondary_date = mydir[17:]
+
+    logging.info("In directory {} looking for file with date {}".format(os.getcwd(), reference_date))
+    reference_file = glob.glob("*%s*.SAFE" % reference_date)[0]
+    secondary_file = glob.glob("*%s*.SAFE" % secondary_date)[0]
+
+    with open("IFM/baseline.log", "r") as f:
+        for line in f:
+            if "estimated baseline perpendicular component" in line:
+                # FIXME: RE is overly complicated here. this is two simple string splits
+                t = re.split(":", line)
+                s = re.split(r'\s+', t[1])
+                baseline = float(s[1])
+
+    back = os.getcwd()
+    os.chdir(os.path.join(reference_file, "annotation"))
+
+    utctime = None
+    for myfile in os.listdir("."):
+        if "001.xml" in myfile:
+            root = etree.parse(myfile)
+            for coord in root.iter('productFirstLineUtcTime'):
+                utc = coord.text
+                logging.info("Found utc time {}".format(utc))
+                t = utc.split("T")
+                logging.info("{}".format(t))
+                s = t[1].split(":")
+                logging.info("{}".format(s))
+                utctime = ((int(s[0]) * 60 + int(s[1])) * 60) + float(s[2])
+    os.chdir(back)
+
+    heading = None
+    name = "IFM/" + reference_date[:8] + ".mli.par"
+    with open(name, "r") as f:
+        for line in f:
+            if "heading" in line:
+                t = re.split(":", line)
+                # FIXME: RE is overly complicated here. this is two simple string splits
+                s = re.split(r'\s+', t[1])
+                heading = float(s[1])
+
+    reference_file = reference_file.replace(".SAFE", "")
+    secondary_file = secondary_file.replace(".SAFE", "")
+
+    os.chdir("PRODUCT")
+    name = "%s.txt" % mydir
+    with open(name, 'w') as f:
+        f.write('Reference Granule: %s\n' % reference_file)
+        f.write('Secondary Granule: %s\n' % secondary_file)
+        f.write('Baseline: %s\n' % baseline)
+        f.write('UTCtime: %s\n' % utctime)
+        f.write('Heading: %s\n' % heading)
+        f.write('Range looks: %s\n' % rlooks)
+        f.write('Azimuth looks: %s\n' % alooks)
+        f.write('INSAR phase filter:  adf\n')
+        f.write('Phase filter parameter: 0.6\n')
+        f.write('Resolution of output (m): %s\n' % res)
+        f.write('Range bandpass filter: no\n')
+        f.write('Azimuth bandpass filter: no\n')
+        f.write('DEM source: %s\n' % dem_source)
+        f.write('DEM resolution (m): %s\n' % (res * 2))
+        f.write('Unwrapping type: mcf\n')
+        f.write('Unwrapping threshold: none\n')
+        f.write('Speckle filtering: off\n')
+    os.chdir("..")
 
 
 def gammaProcess(reference_file, secondary_file, outdir, dem=None, dem_source=None, rlooks=10, alooks=2, inc_flag=False,
